Fix static analysis issues
:Release Notes:
Fix static analysis issues

:Detailed Notes:
Fix static analysis issues

:Testing Performed:
Unit Test Passed

:QA Notes:
NA

:Issues Addressed:
[PLAT-130028] Fix the MISRA defects

+++ src/WebOSLocale.cpp
@@ -85,12 +85,12 @@ void WebOSLocale::parseScript(string &item)
 	if (!script.empty() || item.length() != 4) return;
 
 	bool isValidScript = false;
-	if ( isupper(*item.begin()) )
+	if ( isupper(static_cast<int>(*item.begin())) )
 	{
 		isValidScript = true;
 		for (string::iterator itr = item.begin() + 1; itr != item.end(); ++itr)
 		{
-			if ( isupper(*itr) ) {
+			if ( isupper(static_cast<int>(*itr)) ) {
 				isValidScript = false;
 				break;
 			}
